+++ algorithms/Insertion.cpp
@@ -1,10 +1,11 @@
+#include "Insertion.h"
 #include <algorithm>
 #include <iostream>
 #include <type_traits>
 #include <vector>
 
 
-void insertionSort(std::vector<double> &arr)
+void insertionSort(std::vector<int> &arr)
 {
     if (arr.empty()) {
         return;
