+++ algorithms/Shell.cpp
@@ -1,10 +1,11 @@
+#include "Shell.h"
 #include <algorithm>
 #include <iostream>
 #include <type_traits>
 #include <vector>
 
 
-void shellSort(std::vector<double>& arr, int n) {
+void shellSort(std::vector<int>& arr, int n) {
     for (int gap = n / 2; gap > 0; gap /= 2) {
         for (int i = gap; i < n; i++) {
             int temp = arr[i];
